backend/requirements.txt
Django
mysqlclient
djangorestframework
django-simple-history
django-environ